第二次

+++ antun-rest-app/src/main/java/org/liuyiming/bigdata/antun/web/model/UserAdd.java
@@ -1,105 +1,81 @@
 package org.liuyiming.bigdata.antun.web.model;
 
+/**
+ * 新增数据参数封装类
+ * @author zdjy
+ *
+ */
 public class UserAdd {
-	
-	private String name;
-	
-	private Integer sex;
-	
-	private String birth;
-	
-	private String phone;
-	
-	private String idNo;
-	
-	private Integer province;
-	
-	private Integer city;
-	
-	private Integer town;
-	
-	private String channelCode;
-	
-	private String productCode;
-
+    private String name;
+    private Integer sex;
+    private String birth;
+    private String phone;
+    private String idNo;
+    private Long province;
+    private Long city;
+    private Long town;
+    private String channelCode;
+    private String productCode;
 	public String getName() {
 		return name;
 	}
-
 	public void setName(String name) {
 		this.name = name;
 	}
-
 	public Integer getSex() {
 		return sex;
 	}
-
 	public void setSex(Integer sex) {
 		this.sex = sex;
 	}
-
 	public String getBirth() {
 		return birth;
 	}
-
 	public void setBirth(String birth) {
 		this.birth = birth;
 	}
-
 	public String getPhone() {
 		return phone;
 	}
-
 	public void setPhone(String phone) {
 		this.phone = phone;
 	}
-
 	public String getIdNo() {
 		return idNo;
 	}
-
 	public void setIdNo(String idNo) {
 		this.idNo = idNo;
 	}
-
-	public Integer getProvince() {
+	public Long getProvince() {
 		return province;
 	}
-
-	public void setProvince(Integer province) {
+	public void setProvince(Long province) {
 		this.province = province;
 	}
-
-	public Integer getCity() {
+	public Long getCity() {
 		return city;
 	}
-
-	public void setCity(Integer city) {
+	public void setCity(Long city) {
 		this.city = city;
 	}
-
-	public Integer getTown() {
+	public Long getTown() {
 		return town;
 	}
-
-	public void setTown(Integer town) {
+	public void setTown(Long town) {
 		this.town = town;
 	}
-
 	public String getChannelCode() {
 		return channelCode;
 	}
-
 	public void setChannelCode(String channelCode) {
 		this.channelCode = channelCode;
 	}
-
 	public String getProductCode() {
 		return productCode;
 	}
-
 	public void setProductCode(String productCode) {
 		this.productCode = productCode;
 	}
-
+	
+    
 }
